Move icon resource logic to a new service.
